create updateSchool method to mapping purpose schoolDto ==> school

+++ src/main/java/com/example/final_titv/entity/School.java
@@ -44,7 +44,7 @@ public class School extends BaseEntity{
 //    @JsonIgnoreProperties(value = {"applications", "hibernateLazyInitializer"})
     private Set<TClass> tClassSet;
     @Column(name = "cutoff_score")
-    private Integer cutoffScore; //admission cutoff score
+    private Double cutoffScore; //admission cutoff score
     public void addTClass(TClass tClass){
         try {
             if(tClassSet == null){
